Move collider implementation to engine package
